Module Datetime.Infix

A collection of infix operators for comparing dates.

val (=) : t -> t -> bool

a = b returns true if a equal b, false otherwise.

val (<>) : t -> t -> bool

a <> b returns true if a is not equal to b, false otherwise.

val (>) : t -> t -> bool

a > b returns true if a is greater than b, false otherwise.

val (>=) : t -> t -> bool

a > b returns true if a is greater or equal to b, false otherwise.

val (<) : t -> t -> bool

a > b returns true if a is smaller than b, false otherwise.

val (<=) : t -> t -> bool

a > b returns true if a is smaller or equal to b, false otherwise.