Module Yocaml.Make

A set of functors designed to automate the construction of boring and repetitive modules.

module Runtime (Runtime : Required.RUNTIME) : Required.RUNNER with type 'a t := 'a Eff.t and module Runtime := Runtime

A Runtime is an execution context (ie, Unix or Git). They describe the entry point of a YOCaml program and abstract the file system.

module Data_reader (DP : Required.DATA_PROVIDER) : Required.DATA_READER with type t = DP.t and type 'a eff := 'a Eff.t and type ('a, 'b) arr := ('a, 'b) Task.t and type extraction_strategy := Metadata.extraction_strategy

Builds metadata reader functions based on a data provider.