Module Metadata.Article

include INJECTABLE

type t

val inject : (module Key_value.DESCRIBABLE with type t = 'a) -> t -> (string * 'a) list

Produces a structured object, compliant to a template input from a t.

include READABLE with type t := t

val from_string : (module VALIDABLE) -> string option -> t Validate.t

Try to produces a t from an optional value.

val make : string -> string -> string list -> Date.t -> string option -> string option -> t

val article_title : t -> string

val article_description : t -> string

val date : t -> Date.t

val tags : t -> string list

val title : t -> string option

val description : t -> string option

val set_article_title : string -> t -> t

val set_article_description : string -> t -> t

val set_date : Date.t -> t -> t

val set_tags : string list -> t -> t

val set_title : string option -> t -> t

val set_description : string option -> t -> t

val equal : t -> t -> bool

val pp : Stdlib.Format.formatter -> t -> unit

val compare_by_date : t -> t -> int

val to_rss_item : string -> t -> Rss.Item.t