Module Metadata.Injectable

Parameters

module D : Required.DATA_INJECTABLE

Signature

t = D.t, provides to_data to convert values into Data.t.

include Data.S with type t = D.t

type data := Data.t

Local alias for Yocaml.Data.t.

type t = D.t

The OCaml type that can be converted.

val to_data : t -> data

to_data v converts an OCaml value v of type t into a Yocaml.Data.t.

Re-exports normalize from D to produce metadata records.

include Required.DATA_INJECTABLE with type t := t

val normalize : t -> (string * Data.t) list

Converts a value of type t into a value of type Yocaml.Data.t.