Module Metadata.Page

include INJECTABLE

type t

val inject : (module Key_value.DESCRIBABLE with type t = 'a) -> t -> (string * 'a) list

Produces a structured object, compliant to a template input from a t.

include READABLE with type t := t

val from_string : (module VALIDABLE) -> string option -> t Validate.t

Try to produces a t from an optional value.

val make : string option -> string option -> t

val title : t -> string option

val description : t -> string option

val set_title : string option -> t -> t

val set_description : string option -> t -> t

val equal : t -> t -> bool

val pp : Stdlib.Format.formatter -> t -> unit