Module type Metadata.READABLE

Describes how to transform data processed by the Validable into a concrete type using validation functions from VALIDABLE.

type t

val from_string : (module VALIDABLE) -> string option -> t Validate.t

Try to produces a t from an optional value.