Module type Required.DATA_INJECTABLE

Describes a type capable of being injected as metadata of type Yocaml.Data.t. Used to lift a module into an injecter of arbitrary types to a template (for example).

type t

The type that describes the metadata.

val normalize : t -> (string * Data.t) list

Converts a value of type t into a value of type Yocaml.Data.t.