Module type Required.DATA_READABLE

Describes a type capable of being readable as metadata of type Yocaml.Data.t. Used to lift a module into a validator for data described by a Yocaml.Required.DATA_PROVIDER.

type t

The type that describes the metadata.

val entity_name : string

Assigns a name to an entity (a set of metadata).

val neutral : (t, provider_error) result

Describes a neutral element, as a fallback in the absence of metadata. The function can return an error if the request is mandatory.

val validate : Data.t -> t Data.Validation.validated_value

validate raw_data Validates a data item represented by type Yocaml.Data.t and projects it into a value of type t.