Module Sexp.Canonical

S-Canonical expression used to describe compressed data sources.

Deserialization

val from_string : string -> (t, parsing_error) result

from_string str Try deserializing a string in Csexp.

val from_seq : char Seq.t -> (t, parsing_error) result

from_seq s Try deserializing a sequence of characters in Csexp. The use of a sequence can serve as a basis for easily constructing other sources (string or in_channel for example).

Serialization

val to_buffer : Buffer.t -> t -> unit

to_buffer buf sexp outputs csexp into the given buffer, buf.

val to_string : t -> string

to_string sexp converts a csexp to a string.

val length : t -> int

length sexp gives the length of csexp after serialization.