Module Sexp.Provider

Provider

A node expressed as an S-expression can be used as a Provider.

Dealing with canonical Sexp.

module Canonical : Required.DATA_READER with type t = Sexp.t and type 'a eff := 'a Eff.t and type ('a, 'b) arr := ('a, 'b) Task.t and type extraction_strategy := Metadata.extraction_strategy

Describes a provider for reading Metadata from a DATA_PROVIDER.

type t = Sexp.t

The type that describes the data provider

Reading file with metadata

Just as the Yocaml package describes a low-level interface for propagating effects, the Yocaml.Eff module, and an interface for composing arrows, via the Yocaml.Pipeline module, the plugin describes two sub-modules to serve the same needs.

module Eff : sig ... end

Describes the low-level interface for reading a file and parsing its metadata described by the type t.

module Pipeline : sig ... end

Describes the arrowized interface for reading a file and parsing its metadata.

Data Provider

As it is possible to describe metadata as a Data_provider.

val from_string : string -> (t, Required.provider_error) result

Produces a type t value from a string.

val normalize : t -> Data.t

Converts a value of type t into a value of type Yocaml.Data.t.