Module Task.Dynamic

Utilities for dealing with tasks with dynamic dependencies.

val on_content : ('content_in, 'content_out) t -> (('meta * 'content_in) * Deps.t, ('meta * 'content_out) * Deps.t) t

on_content arr lift an arrow to deal only with the content of a task.

val on_metadata : ('meta_in, 'meta_out) t -> (('meta_in * 'content) * Deps.t, ('meta_out * 'content) * Deps.t) t

on_metadata arr lift an arrow to deal only with the associated metadata of a task.

val on_static : ('meta_in * 'content_in, 'meta_out * 'content_out) t -> (('meta_in * 'content_in) * Deps.t, ('meta_out * 'content_out) * Deps.t) t

on_static arr lift an arrow to deal only with the static part (the couple meta/content).

val on_dependencies : (Deps.t, Deps.t) t -> (('meta * 'content) * Deps.t, ('meta * 'content) * Deps.t) t

on_dependencies arr lift an arrow to deal only with the associated dynamic deps set of a task.

val keep_content : unit -> (('meta * 'content) * Deps.t, 'content * Deps.t) t

keep_content () drop the metadata part of the computed task.

val empty_body : unit -> ('meta * Deps.t, ('meta * string) * Deps.t) t

empty_body () attach an empty body (an empty string) to a task.