Module Yocaml__.Effect

A bad faith preamble

To be beautiful and modern, this project separates the description of the programme from its interpretation. But as the composition is not really to my taste in Preface, I decided to centralize all the effects, like the errors, in one module.

Ugh, that sounds perfectly stupid... it would be like considering that you can only express one family of effects (you could call it ... IO). Don't panic, the first parameter of type effect allows you to make a selective choice when defining Freer. One could say that one takes advantage of the non-surjective aspect of the constructors of a sum (thanks to the GADTs!). Well, I'd be lying if I said I was convinced it was a good approach, but at least it seems viable.

Effects list

Boy, this type sounds like a hell of a lot of trouble to read! don't read it and go a little lower, there are kind of smarts constructors.

type (_, 'a) effects =

| File_exists : Yocaml.Aliases.filepath -> (< file_exists : Yocaml.Aliases.e; .. >, bool) effects

| Get_modification_time : Yocaml.Aliases.filepath -> (< get_modification_time : Yocaml.Aliases.e; .. >, int Yocaml.Try.t) effects

| Read_file : Yocaml.Aliases.filepath -> (< read_file : Yocaml.Aliases.e; .. >, string Yocaml.Try.t) effects

| Write_file : (Yocaml.Aliases.filepath * string) -> (< write_file : Yocaml.Aliases.e; .. >, unit Yocaml.Try.t) effects

| Read_dir : (Yocaml.Aliases.filepath * [< `Files | `Directories | `Both ] * Yocaml.Aliases.filepath Preface.Predicate.t) -> (< read_dir : Yocaml.Aliases.e; .. >, Yocaml.Aliases.filepath list) effects

| Log : (Yocaml.Aliases.log_level * string) -> (< log : Yocaml.Aliases.e; .. >, unit) effects

| Throw : Yocaml.Error.t -> (< throw : Yocaml.Aliases.e; .. >, 'a) effects

| Raise : exn -> (< raise_ : Yocaml.Aliases.e; .. >, 'a) effects

Global definition

Complete mechanism for describing programs by description and providing them with handlers (interpreters/runtime) for all effects modelled in type t. (So absolutely not taking advantage of the slicing capability... It was well worth it!)

Freer monad over effects

All the plumbing for effects description/interpretation resides through a Freer monad (thanks Preface). Although this module is included below, I have taken the liberty of displaying it... for documentation purposes only.

module Freer : Preface_specs.FREER_MONAD with type 'a Freer.f = (< file_exists : Yocaml.Aliases.e; get_modification_time : Yocaml.Aliases.e; read_file : Yocaml.Aliases.e; write_file : Yocaml.Aliases.e; read_dir : Yocaml.Aliases.e; log : Yocaml.Aliases.e; throw : Yocaml.Aliases.e; raise_ : Yocaml.Aliases.e; >, 'a) effects

Performing effects

Once described (or/and specialised), the effects must be produced in a programme description. To transform the description of an effect (a value of type Effect.effect) into the execution of this effect, thus a value of type Effect.t), the perform function is used.

Filesystem

In generating a static blog, having control over the file system seems to be a minimum!

val file_exists : Yocaml.Aliases.filepath -> bool Freer.t

file_exists path should be interpreted as returning true if the file denoted by the file path path exists, false otherwise.

val get_modification_time : Yocaml.Aliases.filepath -> int Yocaml.Try.t Freer.t

get_modification_time path should be interpreted as returning, as an integer, the Unix time (mtime corresponding to the modification date of the file denoted by the file path path.

val read_file : Yocaml.Aliases.filepath -> string Yocaml.Try.t Freer.t

read_file path should be interpreted as trying to read the contents of the file denoted by the file path path. At the moment I'm using strings mainly out of laziness, and as I'll probably be the only user of this library... it doesn't matter!

val write_file : Yocaml.Aliases.filepath -> string -> unit Yocaml.Try.t Freer.t

write_file path content should be interpreted as trying to write content to the file denoted by the file path path. In my understanding of the system, the file will be completely overwritten if it already exists. Once again I am using strings, but this time it is not laziness, it is to be consistent with read_file.

val read_children : Yocaml.Aliases.filepath -> Yocaml.Aliases.filepath Preface.Predicate.t -> Yocaml.Aliases.filepath list Freer.t

Get a list of all children of a path.

val read_child_files : Yocaml.Aliases.filepath -> Yocaml.Aliases.filepath Preface.Predicate.t -> Yocaml.Aliases.filepath list Freer.t

Get a list of all child files of a path (exclude dirs).

val read_child_directories : Yocaml.Aliases.filepath -> Yocaml.Aliases.filepath Preface.Predicate.t -> Yocaml.Aliases.filepath list Freer.t

Get a list of all child directories of a path (exclude files).

val collect_children : Yocaml.Aliases.filepath list -> Yocaml.Aliases.filepath Preface.Predicate.t -> Yocaml.Aliases.filepath list Freer.t

Same of read_children but searching through a list of directories.

val collect_child_files : Yocaml.Aliases.filepath list -> Yocaml.Aliases.filepath Preface.Predicate.t -> Yocaml.Aliases.filepath list Freer.t

Same of read_child_files but searching through a list of directories.

val collect_child_directories : Yocaml.Aliases.filepath list -> Yocaml.Aliases.filepath Preface.Predicate.t -> Yocaml.Aliases.filepath list Freer.t

Same of read_child_directories but searching through a list of directories.

val process_files : Yocaml.Aliases.filepath list -> Yocaml.Aliases.filepath Preface.Predicate.t -> (Yocaml.Aliases.filepath -> unit Freer.t) -> unit Freer.t

process_files path predicate action performs sequentially action on each files which satisfies predicate.

Logging

Even if it would be possible to limit our feedback with the user to simply returning an integer (El famoso Unix Return)... it would still be more convenient to display feedback to the user on the stage the program is in, right?

val log : Yocaml.Aliases.log_level -> string -> unit Freer.t

log level message should be interpreted as writing (probably to standard output) a message associated with a log level. To look good, the colour should change according to the log level, it would look more professional!

val trace : string -> unit Freer.t

trace message is an alias of log Aliases.Trace.

val debug : string -> unit Freer.t

debug message is an alias of log Aliases.Debug.

val info : string -> unit Freer.t

info message is an alias of log Aliases.Info.

val warning : string -> unit Freer.t

warning message is an alias of log Aliases.Warning.

val alert : string -> unit Freer.t

alert message is an alias of log Aliases.Alert.

Open bar

When we are in the context of an IO, ahem, effect execution, it's open bar, we can do whatever we want, like throwing exceptions galore!

val throw : Yocaml.Error.t -> 'a Freer.t

throw error should be interpreted as... "fire, fire, what to do using an Error!".

val raise_ : exn -> 'a Freer.t

raise_ exn should be interpreted as... "fire, fire, what to do using an exception!".

Effects composition

val sequence : 'a list Freer.t -> ('a -> 'b -> 'b Freer.t) -> 'b Freer.t -> 'b Freer.t

Collapses sequentially YOCaml program. sequence ps f p produces a program which performs p followed by f ps. A common usage is p |> sequences ps f.

Included Freer combinators

As mentioned above, the plumbing of program description and program handling is provided through a Freer Monad, a technique that aims to describe a free build over a Left Kan extension. Although the presence of slicing allows for the construction of specialised effects handlers, in the use case of this blog generator, the effects I propagate turn out to be exactly those I have described in my complete effects list. Coicindance, I don't think so!

It therefore seems logical (not to say ergonomic) to introduce the Freer interface in the toplevel of the Effect module. But as the interface is long and tiring to read, I place it at the end of the module!

include Preface_specs.FREER_MONAD with type 'a Effect.f = 'a Freer.f and type 'a Effect.t = 'a Freer.t

type !'a f = 'a Freer.f

type !'a t = 'a Freer.t =

| Return : 'b -> 'b t

| Bind : 'c f * ('c -> 'd t) -> 'd t

type !'a handler = {

handler : b. ('b -> 'a) -> 'b f -> 'a;

}

val perform : 'a f -> 'a t

val run : 'a handler -> 'a t -> 'a

module Functor : sig ... end

module Applicative : sig ... end

module Monad : sig ... end

val bind : ('a -> 'b t) -> 'a t -> 'b t

val map : ('a -> 'b) -> 'a t -> 'b t

val join : 'a t t -> 'a t

val return : 'a -> 'a t

val compose_left_to_right : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

val compose_right_to_left : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

val lift : ('a -> 'b) -> 'a t -> 'b t

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

val replace : 'a -> 'b t -> 'a t

val void : 'a t -> unit t

module Infix : sig ... end

val (=|<) : ('a -> 'b) -> 'a t -> 'b t

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (=<<) : ('a -> 'b t) -> 'a t -> 'b t

val (>=>) : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

val (<=<) : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

val (>>) : unit t -> 'b t -> 'b t

val (<<) : 'a t -> unit t -> 'a t

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

val (<$) : 'a -> 'b t -> 'a t

val ($>) : 'a t -> 'b -> 'b t

module Syntax : sig ... end

val let* : 'a t -> ('a -> 'b t) -> 'b t

val let+ : 'a t -> ('a -> 'b) -> 'b t

module Traverse : Preface.Specs.TRAVERSABLE with type 'a Traverse.t = 'a Freer.t and type 'a Traverse.iter = 'a list