Module Validate.Yaml

val as_object : [> `O of (string * 'a) list ] -> ((string * 'a) list -> 'b) -> 'b -> 'b

as_object yaml valid invalid If yaml is an object, applies the valid function to the key value list, otherwise returns the value invalid.

val fetch_field : string -> (string * 'a) list -> 'a option

fetch_field key list Retrieves the value attached to key in list.

This function is quite suitable for use with as_object:

as_object
  my_object
  (Validate.valid % fetch_field "Hello")
  Error.(to_validate $ Invalid_metadata "object")
(* Returns an option with the field *)

val optional' : (string -> 'a -> 'b option t) -> string -> (string * 'a) list -> 'b option t

Finds an optional field and applies a validation to it. The validator can return None. The function works well with as_object.

val optional : (string -> 'a -> 'b t) -> string -> (string * 'a) list -> 'b option t

Same of optional' except that the validator can never returns an empty value.

val with_default : default:'b -> (string -> 'a -> 'b t) -> string -> (string * 'a) list -> 'b t

Same of optional but handle a default case.

val required : (string -> 'a -> 'b t) -> string -> (string * 'a) list -> 'b t

Same of optional but the field must be present.

val string : string -> [> `String of string ] -> string t

A validator for a String.

val as_string : string -> [> `String of string | `Bool of bool | `Float of float ] -> string t

A validator for a String which can handle bool, float and int as String.

val bool : string -> [> `String of string | `Bool of bool ] -> bool t

A validator for a Boolean.

val int : string -> [> `Float of float ] -> int t

A validator for an Int.

val float : string -> [> `Float of float ] -> float t

A validator for a Float.

val list : (string -> 'a -> 'b t) -> string -> [> `A of 'a list ] -> 'b list t

A validator for List.