Module Yocaml__Metadata.Rules

val is_object : [> `O of 'a ] -> ('a -> 'b) -> 'b -> 'b

is_object obj f x performs f on the values of obj if obj is a JSON Object, otherwise returns x.

val fetch_field : (string * 'a) list -> string -> 'a option

Find a value indexed by a string.

val optional_string : (string * [> `String of string ]) list -> string -> string option Yocaml.Validate.t

Find an optional string in an indexed list.

val optional_date : (string * [> `String of string ]) list -> string -> Date.t option Yocaml.Validate.t

Find an optional date in an indexed list.

val optional_string_list : (string * [> `A of [> `String of string ] list ]) list -> string -> string list Yocaml.Validate.t

Find an optional list of string in an indexed list.

val required_string : (string * [> `String of string ]) list -> string -> string Yocaml.Validate.t

Find a string in an indexed list.

val required_date : (string * [> `String of string ]) list -> string -> Date.t Yocaml.Validate.t

Find a date in an indexed list.

val required_string_list : (string * [> `A of [> `String of string ] list ]) list -> string -> string list Yocaml.Validate.t

Find a nonempty list of string in an indexed list.