Module Yocaml__Util

An invasive but probably useful tooling.

Working with file name

As Wordpress is mainly used to move files, having some functions to work with file names can be a very good idea!

val into : Yocaml.Aliases.filepath -> Yocaml.Aliases.filepath -> Yocaml.Aliases.filepath

filepath |> into dir describes a filepath into a dir.

val with_extension : string -> Yocaml.Aliases.filepath -> bool

with_extension ext path returns true if path ends with ext false otherwise, ie: with_extension "html" "index.html" returns true but with_extenstion "html" "foohtml" returns false.

val basename : Yocaml.Aliases.filepath -> Yocaml.Aliases.filepath

Keep the filename and remove the path.

val add_extension : Yocaml.Aliases.filepath -> string -> Yocaml.Aliases.filepath

Add an extension to a filepath. For example: add_extension "index.txt" "html" will produce "index.txt.html".

val remove_extension : Yocaml.Aliases.filepath -> Yocaml.Aliases.filepath

Remove the extension of a filepath.

val replace_extension : Yocaml.Aliases.filepath -> string -> Yocaml.Aliases.filepath

Replace the extension of a filepath.

String util

As I was not very serious... strings occupy a very large place in Wordpress... so it is necessary to be able to work correctly with them.

val split_metadata : string -> string option * string

Infix operators

Even if sometimes, infix operators can seem unreadable... the immoderate use of Arrows has already made the code incomprehensible... so why deprive yourself?

val ($) : ('a -> 'b) -> 'a -> 'b

f $ x is f @@ x which is f x... but I don't like @@.