Module Yocaml

Here is the list of modules exposed by the YOCaml library. The concept behind this project is to offer, as a library, a set of tools to build pages. In other words, the end user, in this case me, would only have to build a new project in which Preface and YOCaml would be dependencies and then easily create his own static blog generator. It's likely that it won't be efficient or ergonomic, but it's a fun project to do in your spare time.

Please refer to the documentation index for an example.

Build system

Build is the main module of YOCaml. It is used to describe rules attached to dependencies. A static site generator is a collection of ordered rules. (So it is probably not useful to use this project and it would be better to write everything with make, sed and awk).

module Build : sig ... end

Description of a build rule.

module Deps : sig ... end

Deps is an attempt to represent (without much effort) a set of dependencies to build a file.

Handling

Effects Handling

In order to take advantage of Preface (for fun and profit) YOCaml describes a list of effects to manage. As for errors, executable effects are centralised.

module Effect : sig ... end

Centralization of the effects that can be performed.

Errors Handling

Errors handling is mainly based on a biased version of Result and Validation offered by Preface.

module Error : sig ... end

Describes the list of errors that can occur in the process of creating a blog.

module Try : sig ... end

A specialised version of Result (with Error.t as Error part). Try is very useful for producing sequential validations, where as soon as a step produces an error, the computation sequence is interrupted.

module Validate : sig ... end

A specialised version of Validation (with a nonempty list of Error.t as Invalid part). Validate is very useful to produce parallel validations, where each error is accumulated, as opposed to Try which stops the computation at the first error.

Runtime

A YOCaml program is a pure application that has, a priori, no dependence on the operating system. These dependencies are provided when a program is run (as declared in the yocaml_unix package).

module Runtime : sig ... end

Describes a dedicated runtime for varying the execution context of YOCaml.

Misc

Modules serving only internal interests. They are documented (and public) only for the purpose of clarifying the documentation.

module Lexicon : sig ... end

A centralisation of feedback messages.

module Aliases : sig ... end

Some type aliases to clarify the signatures of some functions.

module Util : sig ... end

An invasive but probably useful tooling.

module Metadata : sig ... end

Data structures attachable to articles/documents.

Included general stuff

Included common util

There are always lots of little unreadable tools that I want to use... sometimes it improves readability... sometimes not.

include module type of Util

Working with file name

As Wordpress is mainly used to move files, having some functions to work with file names can be a very good idea!

val into : Aliases.filepath -> Aliases.filepath -> Aliases.filepath

filepath |> into dir describes a filepath into a dir.

val with_extension : string -> Aliases.filepath -> bool

with_extension ext path returns true if path ends with ext false otherwise, ie: with_extension "html" "index.html" returns true but with_extenstion "html" "foohtml" returns false.

val basename : Aliases.filepath -> Aliases.filepath

Keep the filename and remove the path.

val add_extension : Aliases.filepath -> string -> Aliases.filepath

Add an extension to a filepath. For example: add_extension "index.txt" "html" will produce "index.txt.html".

val remove_extension : Aliases.filepath -> Aliases.filepath

Remove the extension of a filepath.

val replace_extension : Aliases.filepath -> string -> Aliases.filepath

Replace the extension of a filepath.

String util

As I was not very serious... strings occupy a very large place in Wordpress... so it is necessary to be able to work correctly with them.

val split_metadata : string -> string option * string

Infix operators

Even if sometimes, infix operators can seem unreadable... the immoderate use of Arrows has already made the code incomprehensible... so why deprive yourself?

val ($) : ('a -> 'b) -> 'a -> 'b

f $ x is f @@ x which is f x... but I don't like @@.

Included Effect plumbery

A page generation process usually involves composing and executing effects, so rather than constantly forcing the Effect module into user space, the module is injected into the high-level API.

include module type of Effect

A bad faith preamble

To be beautiful and modern, this project separates the description of the programme from its interpretation. But as the composition is not really to my taste in Preface, I decided to centralize all the effects, like the errors, in one module.

Ugh, that sounds perfectly stupid... it would be like considering that you can only express one family of effects (you could call it ... IO). Don't panic, the first parameter of type effect allows you to make a selective choice when defining Freer. One could say that one takes advantage of the non-surjective aspect of the constructors of a sum (thanks to the GADTs!). Well, I'd be lying if I said I was convinced it was a good approach, but at least it seems viable.

Effects list

Boy, this type sounds like a hell of a lot of trouble to read! don't read it and go a little lower, there are kind of smarts constructors.

type (_, 'a) effects =

| File_exists : Aliases.filepath -> (< file_exists : Aliases.e; .. >, bool) effects

| Get_modification_time : Aliases.filepath -> (< get_modification_time : Aliases.e; .. >, int Try.t) effects

| Read_file : Aliases.filepath -> (< read_file : Aliases.e; .. >, string Try.t) effects

| Write_file : (Aliases.filepath * string) -> (< write_file : Aliases.e; .. >, unit Try.t) effects

| Read_dir : (Aliases.filepath * [< `Files | `Directories | `Both ] * Aliases.filepath Preface.Predicate.t) -> (< read_dir : Aliases.e; .. >, Aliases.filepath list) effects

| Log : (Aliases.log_level * string) -> (< log : Aliases.e; .. >, unit) effects

| Throw : Error.t -> (< throw : Aliases.e; .. >, 'a) effects

| Raise : exn -> (< raise_ : Aliases.e; .. >, 'a) effects

Global definition

Complete mechanism for describing programs by description and providing them with handlers (interpreters/runtime) for all effects modelled in type t. (So absolutely not taking advantage of the slicing capability... It was well worth it!)

Freer monad over effects

All the plumbing for effects description/interpretation resides through a Freer monad (thanks Preface). Although this module is included below, I have taken the liberty of displaying it... for documentation purposes only.

module Freer = Effect.Freer

Performing effects

Once described (or/and specialised), the effects must be produced in a programme description. To transform the description of an effect (a value of type Effect.effect) into the execution of this effect, thus a value of type Effect.t), the perform function is used.

Filesystem

In generating a static blog, having control over the file system seems to be a minimum!

val file_exists : Aliases.filepath -> bool Freer.t

file_exists path should be interpreted as returning true if the file denoted by the file path path exists, false otherwise.

val get_modification_time : Aliases.filepath -> int Try.t Freer.t

get_modification_time path should be interpreted as returning, as an integer, the Unix time (mtime corresponding to the modification date of the file denoted by the file path path.

val read_file : Aliases.filepath -> string Try.t Freer.t

read_file path should be interpreted as trying to read the contents of the file denoted by the file path path. At the moment I'm using strings mainly out of laziness, and as I'll probably be the only user of this library... it doesn't matter!

val write_file : Aliases.filepath -> string -> unit Try.t Freer.t

write_file path content should be interpreted as trying to write content to the file denoted by the file path path. In my understanding of the system, the file will be completely overwritten if it already exists. Once again I am using strings, but this time it is not laziness, it is to be consistent with read_file.

val read_children : Aliases.filepath -> Aliases.filepath Preface.Predicate.t -> Aliases.filepath list Freer.t

Get a list of all children of a path.

val read_child_files : Aliases.filepath -> Aliases.filepath Preface.Predicate.t -> Aliases.filepath list Freer.t

Get a list of all child files of a path (exclude dirs).

val read_child_directories : Aliases.filepath -> Aliases.filepath Preface.Predicate.t -> Aliases.filepath list Freer.t

Get a list of all child directories of a path (exclude files).

val collect_children : Aliases.filepath list -> Aliases.filepath Preface.Predicate.t -> Aliases.filepath list Freer.t

Same of read_children but searching through a list of directories.

val collect_child_files : Aliases.filepath list -> Aliases.filepath Preface.Predicate.t -> Aliases.filepath list Freer.t

Same of read_child_files but searching through a list of directories.

val collect_child_directories : Aliases.filepath list -> Aliases.filepath Preface.Predicate.t -> Aliases.filepath list Freer.t

Same of read_child_directories but searching through a list of directories.

val process_files : Aliases.filepath list -> Aliases.filepath Preface.Predicate.t -> (Aliases.filepath -> unit Freer.t) -> unit Freer.t

process_files path predicate action performs sequentially action on each files which satisfies predicate.

Logging

Even if it would be possible to limit our feedback with the user to simply returning an integer (El famoso Unix Return)... it would still be more convenient to display feedback to the user on the stage the program is in, right?

val log : Aliases.log_level -> string -> unit Freer.t

log level message should be interpreted as writing (probably to standard output) a message associated with a log level. To look good, the colour should change according to the log level, it would look more professional!

val trace : string -> unit Freer.t

trace message is an alias of log Aliases.Trace.

val debug : string -> unit Freer.t

debug message is an alias of log Aliases.Debug.

val info : string -> unit Freer.t

info message is an alias of log Aliases.Info.

val warning : string -> unit Freer.t

warning message is an alias of log Aliases.Warning.

val alert : string -> unit Freer.t

alert message is an alias of log Aliases.Alert.

Open bar

When we are in the context of an IO, ahem, effect execution, it's open bar, we can do whatever we want, like throwing exceptions galore!

val throw : Error.t -> 'a Freer.t

throw error should be interpreted as... "fire, fire, what to do using an Error!".

val raise_ : exn -> 'a Freer.t

raise_ exn should be interpreted as... "fire, fire, what to do using an exception!".

Effects composition

val sequence : 'a list Freer.t -> ('a -> 'b -> 'b Freer.t) -> 'b Freer.t -> 'b Freer.t

Collapses sequentially YOCaml program. sequence ps f p produces a program which performs p followed by f ps. A common usage is p |> sequences ps f.

Included Freer combinators

As mentioned above, the plumbing of program description and program handling is provided through a Freer Monad, a technique that aims to describe a free build over a Left Kan extension. Although the presence of slicing allows for the construction of specialised effects handlers, in the use case of this blog generator, the effects I propagate turn out to be exactly those I have described in my complete effects list. Coicindance, I don't think so!

It therefore seems logical (not to say ergonomic) to introduce the Freer interface in the toplevel of the Effect module. But as the interface is long and tiring to read, I place it at the end of the module!

include Preface_specs.FREER_MONAD with type 'a Yocaml.f = 'a Freer.f and type 'a Yocaml.t = 'a Freer.t

type !'a f = 'a Freer.f

type !'a t = 'a Freer.t =

| Return : 'b -> 'b t

| Bind : 'c f * ('c -> 'd t) -> 'd t

type !'a handler = {

handler : b. ('b -> 'a) -> 'b f -> 'a;

}

val perform : 'a f -> 'a t

val run : 'a handler -> 'a t -> 'a

module Functor : sig ... end

module Applicative : sig ... end

module Monad : sig ... end

val bind : ('a -> 'b t) -> 'a t -> 'b t

val map : ('a -> 'b) -> 'a t -> 'b t

val join : 'a t t -> 'a t

val return : 'a -> 'a t

val compose_left_to_right : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

val compose_right_to_left : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

val lift : ('a -> 'b) -> 'a t -> 'b t

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

val replace : 'a -> 'b t -> 'a t

val void : 'a t -> unit t

module Infix : sig ... end

val (=|<) : ('a -> 'b) -> 'a t -> 'b t

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (=<<) : ('a -> 'b t) -> 'a t -> 'b t

val (>=>) : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

val (<=<) : ('b -> 'c t) -> ('a -> 'b t) -> 'a -> 'c t

val (>>) : unit t -> 'b t -> 'b t

val (<<) : 'a t -> unit t -> 'a t

val (<$>) : ('a -> 'b) -> 'a t -> 'b t

val (<&>) : 'a t -> ('a -> 'b) -> 'b t

val (<$) : 'a -> 'b t -> 'a t

val ($>) : 'a t -> 'b -> 'b t

module Syntax : sig ... end

val let* : 'a t -> ('a -> 'b t) -> 'b t

val let+ : 'a t -> ('a -> 'b) -> 'b t

module Traverse = Effect.Traverse